requirements.txt
discord.py==1.2.4
captcha==0.3
peewee==3.9.6
